Add upgradeEncodedPassword() functional test.
This is similar to the upgradeFromScrambledPassword() L1 test except it
verifies that the "upgrade" from encoded to encrypted password persists
in the context of Jenkins.  The test could be further expanded as
suggested by the TODOs but for now, being able to poll and confirm only
the encrypted password ends up in the config.xml file is good enough.