Update checkout & history to be changeset-based.
We were already determining the changeset and recording it as part of
the build, this now makes sure we use that same changeset number when
retrieving the files and computing the list of changes since the last
build.

Without these changes, it was possible for the plugin to claim we were
checking out, say, changeset 507, but actually fetch changeset 506,
depending on the exact timing of the changeset and the build's start
time.