Handle switching the user account.
This is a corner case that used to fail when attempting to delete a
workspace owned by the previous user but queried using the current
user's name.  Now we query for all workspaces with the name and attempt
to delete them.  The new user must have the "Administer" permision on
the previous user's workspace to be able to delete it, though.