Simplify to only return an integer.
It's the only piece of information that's actually useful to the
upstream callers and we don't need the Jenkins user name in the log
message, although that's the real reason I had to make this change: the
agents are unable to perform the Jenkins user look-up anyway. On 1.448,
it tries to access the singleton Jenkins instance (NullPointerException)
whereas on 1.620, it is unable to find any CanonicalIdResolver
implementations.