tfs: use owner instead of commiter as author
It is not uncommon to have automatic processes to merge between
branches, run validation before submission, etc. When that happens, the
submitter may be different than the owner (the submitter tends to be an
account used for automation). In those cases, the Jenkins UI will show
the same name for all commits. Use owner name instead, as it better
reflects the name of the person that authored the commit.

Signed-off-by: Alejandro del Castillo <alejandro.delcastillo@ni.com>