Improve VSTS Service Hook integration (#200)
* Reduce code duplication and make PR feedback functional

Deduplicate code paths used for feeding back status to VSTS/TFS.

Further, set the "name" and "genre" fields to sane values:

VSTS uses these fiels for configuration of branch policies. Empty
strings are not permitted, thus this feature was broken on the part of
tfs-plugin.

Use the Jenkins root-URL as 'genre' and the job-name as 'name'.

* Abort running build if a Pull-Request triggers a job

If a Service Hook triggers a job, cancel all running builds with the
same build cause.