TeamPullRequestMergedDetailsAction: persist PullRequest arguments
The TeamPullRequestMergedDetailsAction contains the member
GitPullRequestEx, which in turn contains the details about the
triggering pull request. This member, however is transient, thus
it does not get serialized into the job's build.xml. Since the class
derives something very microsoft-specific, this adds a few wrappers for
serializing the relevant parts of those.

With this, after a restart, Jenkins is able to display which
pull-request actually triggered a build.